% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/pal_functions.R
\name{edi_pal}
\alias{edi_pal}
\title{Interpolates an EDI color palette}
\usage{
edi_pal(palette = "main", reverse = FALSE, ...)
}
\arguments{
\item{palette}{Character name of palette in "edi_palettes" list}

\item{reverse}{Boolean indicating whether the palette should be reversed}

\item{...}{Additional arguments to pass to colorRampPalette() function}
}
\value{
Character of hex codes of the given palette
}
\description{
Function to create a color spectrum of the given palette
}
