% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/pal_functions.R
\name{edi_scale_col}
\alias{edi_scale_col}
\title{Constructs color scale for EDI colors}
\usage{
edi_scale_col(palette = "main", discrete = TRUE, reverse = FALSE, ...)
}
\arguments{
\item{palette}{Character name of palette in "edi_palettes"}

\item{discrete}{Boolean indicating whether color aesthetic is discrete or continuous}

\item{reverse}{Boolean indicating whether the palette should be reversed}

\item{...}{Additional arguments passed to discrete_scale() or
scale_color_gradientn(), used respectively when discrete is TRUE or FALSE}
}
\value{
A ggplot2 color scale based on the inputs
}
\description{
Function to create a color scale of the given palette
}
\examples{
edi_scale_col("mixed", reverse = TRUE)
}
