% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/getPallete.R
\name{get_color_pie}
\alias{get_color_pie}
\title{Color Pie}
\usage{
get_color_pie(number_of_colors)
}
\arguments{
\item{n}{}
}
\value{
plots a color pie, list of color codes (hex format)
}
\description{
This function takes number of colors from user and creates a pie out of it
}
