% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/getPallete.R
\name{get_pallete}
\alias{get_pallete}
\title{Color Pallete}
\usage{
get_pallete(n)
}
\arguments{
\item{n}{- number of distinct colors to be chosen}
}
\value{
list - list of `n` distinct hex values representing colors
}
\description{
The function gives a list of mentioned number of distinct color hex values
}
