% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/getPallete.R
\name{get_pie}
\alias{get_pie}
\title{Create Pie chart}
\usage{
get_pie(l, n)
}
\arguments{
\item{l}{- list of hex values defining the colors}

\item{n}{- number of distinct colors chosen}
}
\value{
pie
}
\description{
This function create a pie from the distinct color returned by the
get_pallete() function
}
